infection error correction

+++ src/main/java/corona/Movement/Adult.java
@@ -249,14 +249,14 @@ public class Adult extends Person {
 
     public void setImmunity() {
         Random r = new Random();
-        immunity = 1 - (5 + r.nextInt(46)) / 100;
+        immunity = 1 - (5 +((double)r.nextInt(46))) / 100;
     }
 
     public void setInfectionRate() throws Exception {
         if (infectionRate < 0) {
             throw new Exception("Invalid infection rate");
         }
-        infectionRate = (1 - (immunity + educationLevel / 500)) * (1 + personScore / 50);
+        infectionRate = (1 - (immunity + (double)educationLevel / 500)) * (1 + (double)personScore / 50);
     }
 
     public Integer getID() {
